Add plex specific configuration
add module to remove comments from the config file so that
it could be parsed properly.

Right now Plex's unique UUID is static, I would like to use the DB
to save this info, and if it changes then update from db. The UUID
needs to be set or else a random one will be chosen and will spam
the plex server