Add tests for saving user config (#13)
* Add initial test for saving user config

Add tests for reading user config

* Add fix for jest debugging with Code and node 8.5

* Make the path argument mandatory

In order to make the functions more pure, and easier to test. I have
removed the deafault path in the user file, and instead it will have
to be passed in manually each time.

With the thinking that config/index.ts file will contain a factory
for creating a user config object which will inject the path.

TODO need to add a function that validates the passed in path variable.

* Add remaining tests for the user config module

Should bring coverage up to 100% for this file.